align inference provider adapters
