--- src/refiner/inference/client.py
@@ -1,7 +1,0 @@
-from __future__ import annotations
-
-from refiner.inference.internal.response import InferenceResponse
-
-__all__ = [
-    "InferenceResponse",
-]
