+++ src/refiner/inference/providers/warnings.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from collections.abc import Collection
+from collections.abc import Collection, Mapping
 
 from refiner.inference.types import InferenceWarning, ProviderOptions
 
@@ -8,15 +8,16 @@ from refiner.inference.types import InferenceWarning, ProviderOptions
 def provider_option_warnings(
     *,
     provider_name: str,
-    expected_namespace: str,
+    expected_namespace: str | Collection[str],
     supported_options: Collection[str],
     provider_options: ProviderOptions | None,
 ) -> list[InferenceWarning]:
     if not provider_options:
         return []
+    expected_namespaces = _expected_namespaces(expected_namespace)
     warnings: list[InferenceWarning] = []
     for namespace in provider_options:
-        if namespace != expected_namespace:
+        if namespace not in expected_namespaces:
             warnings.append(
                 {
                     "type": "unsupported-provider-option",
@@ -27,19 +28,28 @@ def provider_option_warnings(
                     ),
                 }
             )
-    expected_options = provider_options.get(expected_namespace, {})
-    for option in expected_options:
-        if option not in supported_options:
-            warnings.append(
-                {
-                    "type": "unsupported-setting",
-                    "setting": f"providerOptions.{expected_namespace}.{option}",
-                    "message": (
-                        f"{option!r} is not currently mapped by {provider_name}."
-                    ),
-                }
-            )
+    for namespace in expected_namespaces:
+        expected_options = provider_options.get(namespace, {})
+        if isinstance(expected_options, Mapping):
+            for option in expected_options:
+                if option not in supported_options:
+                    warnings.append(
+                        {
+                            "type": "unsupported-setting",
+                            "setting": f"providerOptions.{namespace}.{option}",
+                            "message": (
+                                f"{option!r} is not currently mapped by "
+                                f"{provider_name}."
+                            ),
+                        }
+                    )
     return warnings
+
+
+def _expected_namespaces(namespace: str | Collection[str]) -> tuple[str, ...]:
+    if isinstance(namespace, str):
+        return (namespace,)
+    return tuple(namespace)
 
 
 __all__ = ["provider_option_warnings"]
